backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
coverage==7.13.0
Django==6.0
django-cors-headers==4.9.0
django-defender==0.9.8
django-dotenv==1.4.2
django-ratelimit==4.1.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
factory_boy==3.3.3
Faker==39.0.0
gunicorn==23.0.0
inflection==0.5.1
whitenoise==6.8.2
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
rpds-py==0.30.0
sentry-sdk==2.48.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
